requirements.txt
django-bootstrap5==22.1
mysqlclient
pymysql
google-api-python-client==1.4.1
django-4-jet